# docs/swagger.yaml
openapi: 3.1.0
info:
  title: Api-Symfony-Shell
  version: 1.1.0
servers:
  - url: http://localhost:8000
components:
  securitySchemes:
    bearerCookie:
      type: apiKey
      in: cookie
      name: BEARER
    refreshTokenCookie:
      type: apiKey
      in: cookie
      name: refresh_token
  schemas:
    ErrorResponse:
      type: object
      required: [code, message, details]
      properties:
        code:
          type: string
          examples: [bad_request, unauthorized, access_denied, not_found, conflict, rate_limited, internal_error]
        message:
          type: string
        details:
          nullable: true
          oneOf:
            - type: string
            - type: object
            - type: array
...
    AuditLogDto:
      type: object
      properties:
        id: { type: integer }
        action: { type: string }
        actor: { type: string }
        context: { type: object, additionalProperties: true }
        createdAt: { type: string, format: date-time }
    PaginatedAuditLogs:
      type: object
      required: [data, page, limit, total]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/AuditLogDto'
        page: { type: integer }
        limit: { type: integer }
        total: { type: integer }
paths:
...
  /api/admin/audit-logs:
    get:
      security:
        - bearerCookie: []
      parameters:
        - name: action
          in: query
          schema: { type: string }
        - name: actor
          in: query
          schema: { type: string }
        - name: from
          in: query
          schema: { type: string, format: date-time }
        - name: to
          in: query
          schema: { type: string, format: date-time }
        - name: page
          in: query
          schema: { type: integer, minimum: 1, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 10 }
        - name: sort
          in: query
          schema:
            type: string
            enum: [id, createdAt, action, actor]
            default: id
        - name: order
          in: query
          schema:
            type: string
            enum: [ASC, DESC]
            default: DESC
      responses:
        '200':
          description: Audit list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedAuditLogs'
        '400':
          description: Invalid query parameter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Access denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/admin/audit-logs/export:
    get:
      security:
        - bearerCookie: []
      responses:
        '200':
          description: CSV export audit logs
          content:
            text/csv:
              schema:
                type: string
                format: binary
        '403':
          description: Access denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/admin/audit-logs/{id}:
    get:
      security:
        - bearerCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Audit detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditLogDto'
        '404':
          description: Audit log not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...